backend/handlers/etsy/requirements.txt
feedparser==6.0.11
supabase==2.9.0
python-dotenv==1.0.1
